+++ build-logic-commons/gradle-plugin/build.gradle.kts
@@ -11,6 +11,7 @@ java {
 
 dependencies {
     implementation(project(":code-quality"))
+    implementation(project(":build-scan"))
 
     implementation("org.gradle.kotlin.kotlin-dsl:org.gradle.kotlin.kotlin-dsl.gradle.plugin:2.2.0")
     implementation("org.gradle.kotlin:gradle-kotlin-dsl-conventions:0.7.0")
