+++ build-logic/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+    id("gradlebuild.cache-miss-monitor")
+}
+
 description = "Provides plugins that are used by Gradle subprojects"
 
 tasks.register("check") {
